pyproject.toml
[project]
name = "langgraph-boot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "openai>=1.97.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "selenium>=4.34.2",
    "uvicorn>=0.35.0",
]
